+++ platforms/native/components/Column/index.ts
@@ -1,14 +1,6 @@
 import styled from 'styled-components/native'
-import { ShallotProp, getStyle } from '@shallot-ui/core'
+import { getStyle } from '@shallot-ui/core'
+import { ColumnShallot, withColumnStyleProps } from '@shallot-ui/column'
 
-export const Column = styled.View<{ shallot?: ShallotProp }>(
-  ({ shallot, ...rest }) =>
-    getStyle({
-      shallot: {
-        display: 'flex',
-        flexDirection: 'column',
-        ...shallot,
-      },
-      ...rest,
-    }),
-)
+const RawColumn = styled.View<{ shallot?: ColumnShallot }>(getStyle)
+export const Column = withColumnStyleProps(RawColumn)