+++ platforms/web/components/Column/index.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
-import { ShallotProp, getStyle } from '@shallot-ui/core'
-import { withColumnStyleProps } from '@shallot-ui/column'
+import { getStyle } from '@shallot-ui/core'
+import { withColumnStyleProps, ColumnShallot } from '@shallot-ui/column'
 
-const Base = styled.div<{ shallot?: ShallotProp }>(getStyle)
+const Base = styled.div<{ shallot?: ColumnShallot }>(getStyle)
 export const Column = withColumnStyleProps(Base)